kafka_producer/requirements.txt
requests
confluent-kafka
python-dotenv
